+++ daemon/config/builder.go
@@ -11,9 +11,37 @@ import (
 
 // BuilderGCRule represents a GC rule for buildkit cache
 type BuilderGCRule struct {
-	All         bool            `json:",omitempty"`
-	Filter      BuilderGCFilter `json:",omitempty"`
-	KeepStorage string          `json:",omitempty"`
+	All           bool            `json:",omitempty"`
+	Filter        BuilderGCFilter `json:",omitempty"`
+	ReservedSpace string          `json:",omitempty"`
+	MaxUsedSpace  string          `json:",omitempty"`
+	MinFreeSpace  string          `json:",omitempty"`
+}
+
+func (x *BuilderGCRule) UnmarshalJSON(data []byte) error {
+	var xx struct {
+		All           bool            `json:",omitempty"`
+		Filter        BuilderGCFilter `json:",omitempty"`
+		ReservedSpace string          `json:",omitempty"`
+		MaxUsedSpace  string          `json:",omitempty"`
+		MinFreeSpace  string          `json:",omitempty"`
+
+		// Deprecated option is now equivalent to ReservedSpace.
+		KeepStorage string `json:",omitempty"`
+	}
+	if err := json.Unmarshal(data, &xx); err != nil {
+		return err
+	}
+
+	x.All = xx.All
+	x.Filter = xx.Filter
+	x.ReservedSpace = xx.ReservedSpace
+	x.MaxUsedSpace = xx.MaxUsedSpace
+	x.MinFreeSpace = xx.MinFreeSpace
+	if x.ReservedSpace == "" {
+		x.ReservedSpace = xx.KeepStorage
+	}
+	return nil
 }
 
 // BuilderGCFilter contains garbage-collection filter rules for a BuildKit builder
@@ -56,9 +84,38 @@ func (x *BuilderGCFilter) UnmarshalJSON(data []byte) error {
 
 // BuilderGCConfig contains GC config for a buildkit builder
 type BuilderGCConfig struct {
-	Enabled            bool            `json:",omitempty"`
-	Policy             []BuilderGCRule `json:",omitempty"`
-	DefaultKeepStorage string          `json:",omitempty"`
+	Enabled              bool            `json:",omitempty"`
+	Policy               []BuilderGCRule `json:",omitempty"`
+	DefaultReservedSpace string          `json:",omitempty"`
+	DefaultMaxUsedSpace  string          `json:",omitempty"`
+	DefaultMinFreeSpace  string          `json:",omitempty"`
+}
+
+func (x *BuilderGCConfig) UnmarshalJSON(data []byte) error {
+	var xx struct {
+		Enabled              bool            `json:",omitempty"`
+		Policy               []BuilderGCRule `json:",omitempty"`
+		DefaultReservedSpace string          `json:",omitempty"`
+		DefaultMaxUsedSpace  string          `json:",omitempty"`
+		DefaultMinFreeSpace  string          `json:",omitempty"`
+
+		// Deprecated option is now equivalent to DefaultReservedSpace.
+		DefaultKeepStorage string `json:",omitempty"`
+	}
+
+	if err := json.Unmarshal(data, &xx); err != nil {
+		return err
+	}
+
+	x.Enabled = xx.Enabled
+	x.Policy = xx.Policy
+	x.DefaultReservedSpace = xx.DefaultReservedSpace
+	x.DefaultMaxUsedSpace = xx.DefaultMaxUsedSpace
+	x.DefaultMinFreeSpace = xx.DefaultMinFreeSpace
+	if x.DefaultReservedSpace == "" {
+		x.DefaultReservedSpace = xx.DefaultKeepStorage
+	}
+	return nil
 }
 
 // BuilderHistoryConfig contains history config for a buildkit builder